# generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "node-bill API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/engagements/:id/entities": {
      "get": {
        "operationId": "secu_engagement_entities_list",
        "tags": [
          "secu-engagements"
        ],
        "summary": "List entities linked to an engagement",
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "asset_domain",
                "asset_subdomain",
                "asset_ip",
                "asset_host",
                "asset_url",
                "person",
                "organization",
                "location",
                "credential_ref",
                "document"
              ]
            }
          }
        ],
        "security": [
          {
            "frontendBearer": []
          }
        ],
        "x-transport": "http"
      },
      "post": {
        "operationId": "secu_engagement_entity_link",
        "tags": [
          "secu-engagements"
        ],
        "summary": "Link an entity to an engagement (or upsert+link in one call)",
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "201": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "entityId": {
                    "type": "integer",
                    "exclusiveMinimum": 0
                  },
                  "upsert": {
                    "type": "object",
                    "properties": {
                      "kind": {
                        "type": "string",
                        "enum": [
                          "asset_domain",
                          "asset_subdomain",
                          "asset_ip",
                          "asset_host",
                          "asset_url",
                          "person",
                          "organization",
                          "location",
                          "credential_ref",
                          "document"
                        ]
                      },
                      "primaryValue": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 512
                      },
                      "displayName": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 256
                      },
                      "discriminator": {
                        "anyOf": [
                          {
                            "type": "string",
                            "maxLength": 256
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "data": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    },
                    "required": [
                      "kind",
                      "primaryValue"
                    ],
                    "additionalProperties": false
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "primary_target",
                      "in_scope",
                      "out_of_scope",
                      "pivot",
                      "context"
                    ]
                  },
                  "notes": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 4096
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "security": [
          {
            "frontendBearer": []
          }
        ],
        "x-transport": "http"
      }
    },
    "/engagements/:id/entities/:entityId": {
      "delete": {
        "operationId": "secu_engagement_entity_unlink",
        "tags": [
          "secu-engagements"
        ],
        "summary": "Unlink an entity from an engagement",
        "responses": {
          "204": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            }
          },
          {
            "name": "entityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            }
          }
        ],
        "security": [
          {
            "frontendBearer": []
          }
        ],
        "x-transport": "http"
      }
    },
...
    "/entities": {
      "post": {
        "operationId": "secu_entity_upsert",
        "tags": [
          "secu-entities"
        ],
        "summary": "Create or upsert a global entity (deduplicated via canonical_key)",
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "kind": {
                    "type": "string",
                    "enum": [
                      "asset_domain",
                      "asset_subdomain",
                      "asset_ip",
                      "asset_host",
                      "asset_url",
                      "person",
                      "organization",
                      "location",
                      "credential_ref",
                      "document"
                    ]
                  },
                  "primaryValue": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 512
                  },
                  "displayName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "discriminator": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 256
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "data": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                },
                "required": [
                  "kind",
                  "primaryValue"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "security": [
          {
            "frontendBearer": []
          }
        ],
        "x-transport": "http"
      },
      "get": {
        "operationId": "secu_entity_search",
        "tags": [
          "secu-entities"
        ],
        "summary": "Search global entities (kind + free text)",
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "asset_domain",
                "asset_subdomain",
                "asset_ip",
                "asset_host",
                "asset_url",
                "person",
                "organization",
                "location",
                "credential_ref",
                "document"
              ]
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 256
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "security": [
          {
            "frontendBearer": []
          }
        ],
        "x-transport": "http"
      }
    },
    "/entities/:id": {
      "get": {
        "operationId": "secu_entity_get",
        "tags": [
          "secu-entities"
        ],
        "summary": "Get entity detail (incl. tags, engagements, relationship-count)",
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            }
          }
        ],
        "security": [
          {
            "frontendBearer": []
          }
        ],
        "x-transport": "http"
      }
    },
    "/entities/:id/relationships": {
      "get": {
        "operationId": "secu_entity_relationships_list",
        "tags": [
          "secu-entities"
        ],
        "summary": "List relationships incident to an entity",
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            }
          }
        ],
        "security": [
          {
            "frontendBearer": []
          }
        ],
        "x-transport": "http"
      },
      "post": {
        "operationId": "secu_entity_relationship_upsert",
        "tags": [
          "secu-entities"
        ],
        "summary": "Upsert a relationship from this entity to another",
        "responses": {
          "201": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "toEntityId": {
                    "type": "integer",
                    "exclusiveMinimum": 0
                  },
                  "kind": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 64
                  },
                  "confidence": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 100
                  },
                  "source": {
                    "type": "string",
                    "maxLength": 64
                  },
                  "data": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                },
                "required": [
                  "toEntityId",
                  "kind"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "security": [
          {
            "frontendBearer": []
          }
        ],
        "x-transport": "http"
      }
    },
    "/entities/:id/tags": {
      "post": {
        "operationId": "secu_entity_tag_add",
        "tags": [
          "secu-entities"
        ],
        "summary": "Add a tag to an entity (idempotent)",
        "responses": {
          "201": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tag": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 64
                  },
                  "color": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 16
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "required": [
                  "tag"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "security": [
          {
            "frontendBearer": []
          }
        ],
        "x-transport": "http"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "xApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "cronJobBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "CRON_SECRET"
      },
      "frontendBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "oauth2Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "companyApiKeyBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_KEY"
      }
    }
  }
}